Respond to reviews, move refactored algs back to main directories and not experimental
