+++ python/pylibcugraph/pylibcugraph/experimental/__init__.py
@@ -68,6 +68,3 @@ node2vec = promoted_experimental_warning_wrapper(node2vec)
 
 from pylibcugraph.uniform_neighborhood_sampling import EXPERIMENTAL__uniform_neighborhood_sampling
 uniform_neighborhood_sampling = experimental_warning_wrapper(EXPERIMENTAL__uniform_neighborhood_sampling)
-
-from pylibcugraph.katz_centrality import EXPERIMENTAL__katz_centrality
-katz_centrality = experimental_warning_wrapper(EXPERIMENTAL__katz_centrality)
